decoupled testing the redis db accessor with the method checking it's connection - added factory to pass concrete implementations rather than directly accessing method
